Changed the package structure. Moved things to be a bit more clear.
