+++ src/net/dv8tion/discord/commands/MyAnimeListCommand.java
@@ -3,8 +3,8 @@ package net.dv8tion.discord.commands;
 import java.util.Arrays;
 import java.util.List;
 
-import net.dv8tion.discord.Downloader;
-import net.dv8tion.discord.GoogleSearch;
+import net.dv8tion.discord.util.Downloader;
+import net.dv8tion.discord.util.GoogleSearch;
 
 import org.apache.commons.lang3.StringUtils;
