+++ src/net/dv8tion/discord/commands/SearchCommand.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import me.itsghost.jdiscord.events.UserChatEvent;
 import me.itsghost.jdiscord.message.MessageBuilder;
-import net.dv8tion.discord.GoogleSearch;
+import net.dv8tion.discord.util.GoogleSearch;
 
 import org.apache.commons.lang3.StringUtils;
