+++ src/net/dv8tion/discord/util/Downloader.java
@@ -1,4 +1,4 @@
-package net.dv8tion.discord;
+package net.dv8tion.discord.util;
 import java.awt.image.BufferedImage;
 import java.io.BufferedInputStream;
 import java.io.BufferedReader;
